Hw5

+++ scrapper/src/main/java/edu/java/model/scrapper/dto/request/LinkUpdateRequest.java
@@ -5,13 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 public class LinkUpdateRequest extends MyRequest {
-    private long id;
+    private Long id;
     private String url;
     private String description;
-    private List<Integer> tgChatIds;
+    private List<Long> tgChatIds;
 }
